requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
geopy==2.4.1
django-redis>=5.4.0,<6.0.0
polyline>=2.0.0,<3.0.0